docs(viewer-docs): add mock data for line mode
