+++ apps/insight-viewer-docs/mocks/lines.ts
@@ -1,420 +1,104 @@
 /* eslint-disable import/no-unresolved */
 import { Annotation } from '@lunit/insight-viewer'
 
-export function labelFunction(annotation: Annotation): string {
-  return `Annotation(${annotation.id})`
-}
-
-export const categoryColors = {
-  normal: '#3366cc',
-  abnormal: '#dc3912',
-} as const
-
 export const initialAnnotation: Annotation[] = [
   {
     id: 1,
     polygon: [
-      [771.1257142857143, 552.77],
-      [771.1257142857143, 548.9528571428572],
-      [774.9428571428572, 548.9528571428572],
-      [778.76, 548.9528571428572],
-      [782.5771428571429, 545.1357142857142],
-      [790.2114285714285, 545.1357142857142],
-      [797.8457142857143, 541.3185714285714],
-      [805.48, 537.5014285714286],
-      [820.7485714285714, 529.8671428571429],
-      [836.0171428571429, 526.05],
-      [855.1028571428571, 522.2328571428571],
-      [870.3714285714286, 514.5985714285714],
-      [889.4571428571429, 510.7814285714286],
-      [916.1771428571428, 506.9642857142857],
-      [931.4457142857143, 503.14714285714285],
-      [946.7142857142858, 503.14714285714285],
-      [954.3485714285714, 503.14714285714285],
-      [969.6171428571429, 510.7814285714286],
-      [981.0685714285714, 514.5985714285714],
-      [996.3371428571429, 526.05],
-      [1007.7885714285715, 537.5014285714286],
-      [1026.8742857142856, 552.77],
-      [1038.3257142857142, 568.0385714285715],
-      [1053.5942857142857, 590.9414285714286],
-      [1065.0457142857142, 621.4785714285714],
-      [1076.4971428571428, 652.0157142857142],
-      [1076.4971428571428, 671.1014285714285],
-      [1080.3142857142857, 701.6385714285715],
-      [1084.1314285714286, 728.3585714285714],
-      [1087.9485714285715, 747.4442857142857],
-      [1087.9485714285715, 785.6157142857143],
-      [1087.9485714285715, 816.1528571428571],
-      [1095.5828571428572, 842.8728571428571],
-      [1095.5828571428572, 869.5928571428572],
-      [1095.5828571428572, 896.3128571428572],
-      [1095.5828571428572, 923.0328571428571],
-      [1095.5828571428572, 945.9357142857143],
-      [1091.7657142857142, 968.8385714285714],
-      [1087.9485714285715, 991.7414285714286],
-      [1072.68, 1018.4614285714285],
-      [1053.5942857142857, 1048.9985714285715],
-      [1030.6914285714286, 1075.7185714285715],
-      [1000.1542857142858, 1106.2557142857142],
-      [981.0685714285714, 1121.5242857142857],
-      [958.1657142857143, 1140.61],
-      [935.2628571428572, 1159.6957142857143],
-      [904.7257142857143, 1174.9642857142858],
-      [881.8228571428572, 1186.4157142857143],
-      [855.1028571428571, 1201.6842857142858],
-      [824.5657142857143, 1213.1357142857144],
-      [794.0285714285715, 1224.587142857143],
-      [755.8571428571429, 1232.2214285714285],
-      [732.9542857142857, 1239.8557142857144],
-      [698.6, 1243.672857142857],
-      [668.0628571428572, 1243.672857142857],
-      [641.3428571428572, 1243.672857142857],
-      [626.0742857142857, 1243.672857142857],
-      [618.44, 1239.8557142857144],
-      [618.44, 1239.8557142857144],
-      [614.6228571428571, 1232.2214285714285],
-      [603.1714285714286, 1228.4042857142858],
-      [599.3542857142858, 1220.77],
-      [587.9028571428571, 1213.1357142857144],
-      [572.6342857142857, 1197.8671428571429],
-      [572.6342857142857, 1194.05],
-      [565, 1186.4157142857143],
-      [557.3657142857143, 1174.9642857142858],
-      [549.7314285714285, 1155.8785714285714],
-      [545.9142857142857, 1140.61],
-      [538.28, 1125.3414285714287],
-      [526.8285714285714, 1098.6214285714286],
-      [519.1942857142857, 1083.3528571428571],
-      [515.3771428571429, 1071.9014285714286],
-      [511.56, 1056.6328571428571],
-      [507.7428571428571, 1045.1814285714286],
-      [507.7428571428571, 1029.912857142857],
-      [503.92571428571426, 1018.4614285714285],
-      [503.92571428571426, 1010.8271428571429],
-      [503.92571428571426, 1003.1928571428572],
-      [503.92571428571426, 991.7414285714286],
-      [507.7428571428571, 976.4728571428572],
-      [515.3771428571429, 961.2042857142857],
-      [523.0114285714286, 945.9357142857143],
-      [530.6457142857142, 934.4842857142858],
-      [538.28, 923.0328571428571],
-      [545.9142857142857, 915.3985714285715],
-      [557.3657142857143, 900.13],
-      [572.6342857142857, 888.6785714285714],
-      [580.2685714285715, 884.8614285714286],
-      [587.9028571428571, 881.0442857142857],
-      [595.5371428571428, 869.5928571428572],
-      [614.6228571428571, 858.1414285714286],
-      [626.0742857142857, 850.5071428571429],
-      [637.5257142857143, 846.69],
-      [641.3428571428572, 842.8728571428571],
-      [641.3428571428572, 842.8728571428571],
-      [641.3428571428572, 842.8728571428571],
-      [645.16, 842.8728571428571],
-      [645.16, 839.0557142857143],
-      [645.16, 839.0557142857143],
-      [645.16, 839.0557142857143],
-      [648.9771428571429, 839.0557142857143],
-      [648.9771428571429, 839.0557142857143],
-      [648.9771428571429, 839.0557142857143],
-      [652.7942857142857, 835.2385714285714],
-      [652.7942857142857, 835.2385714285714],
-      [656.6114285714286, 835.2385714285714],
-      [656.6114285714286, 835.2385714285714],
-      [656.6114285714286, 831.4214285714286],
-      [656.6114285714286, 831.4214285714286],
-      [660.4285714285714, 831.4214285714286],
+      [164.19999999999993, 510.7814285714286],
+      [91.67428571428566, 850.5071428571429],
     ],
-    dataAttrs: {
-      'data-category': 'normal',
-    },
+    labelPosition: [164.19999999999993, 510.7814285714286],
+    lineWidth: 1.5,
   },
   {
     id: 2,
     polygon: [
-      [1729.2285714285715, 865.7757142857143],
-      [1733.0457142857142, 861.9585714285714],
-      [1733.0457142857142, 858.1414285714286],
-      [1744.4971428571428, 854.3242857142857],
-      [1759.7657142857142, 842.8728571428571],
-      [1778.8514285714286, 835.2385714285714],
-      [1805.5714285714287, 823.7871428571428],
-      [1828.4742857142858, 812.3357142857143],
-      [1862.8285714285714, 800.8842857142857],
-      [1904.817142857143, 793.25],
-      [1935.3542857142857, 785.6157142857143],
-      [1954.44, 781.7985714285714],
-      [1958.2571428571428, 781.7985714285714],
-      [1962.074285714286, 781.7985714285714],
-      [1962.074285714286, 781.7985714285714],
-      [1962.074285714286, 785.6157142857143],
-      [1965.8914285714286, 785.6157142857143],
-      [1965.8914285714286, 793.25],
-      [1973.5257142857145, 800.8842857142857],
-      [1977.3428571428572, 808.5185714285715],
-      [1981.1599999999999, 823.7871428571428],
-      [1992.6114285714284, 846.69],
-      [1996.4285714285716, 865.7757142857143],
-      [2004.062857142857, 892.4957142857143],
-      [2004.062857142857, 930.6671428571428],
-      [2011.6971428571428, 957.3871428571429],
-      [2015.5142857142855, 980.29],
-      [2019.3314285714287, 1010.8271428571429],
-      [2019.3314285714287, 1029.912857142857],
-      [2019.3314285714287, 1056.6328571428571],
-      [2019.3314285714287, 1079.5357142857142],
-      [2019.3314285714287, 1098.6214285714286],
-      [2007.88, 1117.7071428571428],
-      [1996.4285714285716, 1140.61],
-      [1981.1599999999999, 1155.8785714285714],
-      [1969.7085714285713, 1171.1471428571429],
-      [1954.44, 1178.7814285714285],
-      [1942.9885714285715, 1186.4157142857143],
-      [1935.3542857142857, 1190.2328571428573],
-      [1923.902857142857, 1194.05],
-      [1912.4514285714286, 1194.05],
-      [1897.182857142857, 1194.05],
-      [1885.7314285714285, 1194.05],
-      [1874.28, 1197.8671428571429],
-      [1862.8285714285714, 1197.8671428571429],
-      [1855.1942857142858, 1197.8671428571429],
-      [1855.1942857142858, 1201.6842857142858],
-      [1855.1942857142858, 1201.6842857142858],
-      [1851.3771428571429, 1201.6842857142858],
-      [1851.3771428571429, 1197.8671428571429],
-      [1851.3771428571429, 1197.8671428571429],
-      [1851.3771428571429, 1194.05],
-      [1847.56, 1186.4157142857143],
-      [1836.1085714285714, 1174.9642857142858],
-      [1828.4742857142858, 1159.6957142857143],
-      [1820.84, 1144.4271428571428],
-      [1801.7542857142857, 1121.5242857142857],
-      [1794.1200000000001, 1102.4385714285715],
-      [1782.6685714285713, 1083.3528571428571],
-      [1778.8514285714286, 1068.0842857142857],
-      [1771.2171428571428, 1052.8157142857142],
-      [1767.4, 1037.547142857143],
-      [1763.5828571428572, 1022.2785714285715],
-      [1763.5828571428572, 1014.6442857142857],
-      [1759.7657142857142, 1010.8271428571429],
-      [1759.7657142857142, 1007.01],
-      [1759.7657142857142, 1007.01],
-      [1759.7657142857142, 1007.01],
-      [1759.7657142857142, 1007.01],
+      [244.35999999999996, 1159.6957142857143],
+      [400.8628571428571, 1629.2042857142858],
     ],
-    dataAttrs: {
-      'data-category': 'abnormal',
-    },
+    labelPosition: [244.35999999999996, 1159.6957142857143],
+    lineWidth: 1.5,
   },
   {
     id: 3,
     polygon: [
-      [1133.7542857142857, 1537.5928571428572],
-      [1133.7542857142857, 1537.5928571428572],
-      [1129.9371428571428, 1537.5928571428572],
-      [1126.1200000000001, 1541.41],
-      [1118.4857142857143, 1545.2271428571428],
-      [1110.8514285714286, 1549.0442857142857],
-      [1099.4, 1556.6785714285713],
-      [1091.7657142857142, 1564.3128571428572],
-      [1084.1314285714286, 1575.7642857142857],
-      [1072.68, 1587.2157142857143],
-      [1065.0457142857142, 1602.4842857142858],
-      [1061.2285714285715, 1610.1185714285714],
-      [1049.777142857143, 1625.3871428571429],
-      [1042.142857142857, 1652.107142857143],
-      [1038.3257142857142, 1667.3757142857144],
-      [1038.3257142857142, 1686.4614285714285],
-      [1038.3257142857142, 1701.73],
-      [1042.142857142857, 1716.9985714285715],
-      [1053.5942857142857, 1732.267142857143],
-      [1057.4114285714286, 1739.9014285714286],
-      [1076.4971428571428, 1755.17],
-      [1091.7657142857142, 1762.8042857142857],
-      [1110.8514285714286, 1770.4385714285715],
-      [1129.9371428571428, 1774.2557142857142],
-      [1152.84, 1778.0728571428572],
-      [1168.1085714285714, 1781.89],
-      [1183.3771428571429, 1781.89],
-      [1198.6457142857143, 1781.89],
-      [1210.097142857143, 1778.0728571428572],
-      [1217.7314285714285, 1774.2557142857142],
-      [1217.7314285714285, 1774.2557142857142],
-      [1221.5485714285714, 1770.4385714285715],
-      [1221.5485714285714, 1766.6214285714286],
-      [1221.5485714285714, 1762.8042857142857],
-      [1225.3657142857144, 1758.9871428571428],
-      [1229.182857142857, 1747.5357142857142],
-      [1229.182857142857, 1739.9014285714286],
-      [1229.182857142857, 1732.267142857143],
-      [1229.182857142857, 1716.9985714285715],
-      [1229.182857142857, 1705.547142857143],
-      [1225.3657142857144, 1682.6442857142856],
-      [1225.3657142857144, 1671.192857142857],
-      [1221.5485714285714, 1652.107142857143],
-      [1221.5485714285714, 1633.0214285714285],
-      [1213.9142857142858, 1617.7528571428572],
-      [1210.097142857143, 1598.6671428571428],
-      [1202.462857142857, 1575.7642857142857],
-      [1194.8285714285714, 1564.3128571428572],
-      [1183.3771428571429, 1552.8614285714286],
-      [1164.2914285714285, 1541.41],
-      [1152.84, 1537.5928571428572],
-      [1133.7542857142857, 1533.7757142857142],
-      [1122.3028571428572, 1533.7757142857142],
-      [1110.8514285714286, 1529.9585714285715],
-      [1110.8514285714286, 1529.9585714285715],
-      [1107.0342857142857, 1529.9585714285715],
-      [1103.2171428571428, 1529.9585714285715],
-      [1103.2171428571428, 1529.9585714285715],
+      [377.9599999999999, 1900.2214285714285],
+      [855.1028571428571, 1835.33],
     ],
-    dataAttrs: {
-      'data-category': 'normal',
-    },
+    labelPosition: [377.9599999999999, 1900.2214285714285],
+    lineWidth: 1.5,
   },
   {
     id: 4,
     polygon: [
-      [1992.6114285714284, 1694.0957142857144],
-      [1992.6114285714284, 1690.2785714285715],
-      [1992.6114285714284, 1690.2785714285715],
-      [1988.7942857142857, 1690.2785714285715],
-      [1988.7942857142857, 1690.2785714285715],
-      [1977.3428571428572, 1694.0957142857144],
-      [1965.8914285714286, 1705.547142857143],
-      [1950.6228571428574, 1713.1814285714286],
-      [1916.2685714285715, 1739.9014285714286],
-      [1904.817142857143, 1751.3528571428571],
-      [1889.5485714285714, 1770.4385714285715],
-      [1881.9142857142858, 1785.7071428571428],
-      [1878.097142857143, 1789.5242857142857],
-      [1878.097142857143, 1793.3414285714287],
-      [1878.097142857143, 1793.3414285714287],
-      [1878.097142857143, 1793.3414285714287],
-      [1881.9142857142858, 1793.3414285714287],
-      [1889.5485714285714, 1797.1585714285713],
-      [1897.182857142857, 1797.1585714285713],
-      [1912.4514285714286, 1797.1585714285713],
-      [1927.72, 1793.3414285714287],
-      [1942.9885714285715, 1789.5242857142857],
-      [1958.2571428571428, 1778.0728571428572],
-      [1958.2571428571428, 1778.0728571428572],
-      [1969.7085714285713, 1762.8042857142857],
-      [1969.7085714285713, 1758.9871428571428],
-      [1973.5257142857145, 1755.17],
-      [1973.5257142857145, 1755.17],
-      [1973.5257142857145, 1755.17],
-      [1973.5257142857145, 1751.3528571428571],
-      [1973.5257142857145, 1751.3528571428571],
-      [1973.5257142857145, 1751.3528571428571],
-      [1973.5257142857145, 1747.5357142857142],
-      [1973.5257142857145, 1747.5357142857142],
-      [1973.5257142857145, 1747.5357142857142],
-      [1973.5257142857145, 1743.7185714285715],
-      [1973.5257142857145, 1743.7185714285715],
-      [1973.5257142857145, 1743.7185714285715],
-      [1973.5257142857145, 1743.7185714285715],
+      [973.4342857142858, 2308.655714285714],
+      [1591.8114285714287, 1850.5985714285714],
     ],
-    dataAttrs: {
-      'data-category': 'normal',
-    },
+    labelPosition: [973.4342857142858, 2308.655714285714],
+    lineWidth: 1.5,
   },
   {
     id: 5,
     polygon: [
-      [1599.4457142857143, 1476.5185714285715],
-      [1599.4457142857143, 1476.5185714285715],
-      [1595.6285714285714, 1476.5185714285715],
-      [1595.6285714285714, 1476.5185714285715],
-      [1595.6285714285714, 1480.3357142857144],
-      [1595.6285714285714, 1480.3357142857144],
-      [1591.8114285714287, 1480.3357142857144],
-      [1591.8114285714287, 1484.152857142857],
-      [1591.8114285714287, 1484.152857142857],
-      [1591.8114285714287, 1484.152857142857],
-      [1591.8114285714287, 1487.97],
-      [1595.6285714285714, 1487.97],
-      [1595.6285714285714, 1487.97],
-      [1599.4457142857143, 1491.787142857143],
-      [1603.2628571428572, 1491.787142857143],
-      [1610.8971428571429, 1491.787142857143],
-      [1614.7142857142858, 1495.6042857142857],
-      [1622.3485714285714, 1495.6042857142857],
-      [1626.1657142857143, 1495.6042857142857],
-      [1626.1657142857143, 1495.6042857142857],
-      [1629.9828571428573, 1495.6042857142857],
-      [1629.9828571428573, 1495.6042857142857],
-      [1629.9828571428573, 1495.6042857142857],
-      [1629.9828571428573, 1491.787142857143],
-      [1629.9828571428573, 1491.787142857143],
-      [1629.9828571428573, 1491.787142857143],
-      [1633.8, 1491.787142857143],
-      [1633.8, 1487.97],
-      [1633.8, 1487.97],
-      [1633.8, 1487.97],
-      [1633.8, 1487.97],
-      [1633.8, 1484.152857142857],
-      [1633.8, 1484.152857142857],
-      [1633.8, 1484.152857142857],
-      [1633.8, 1480.3357142857144],
-      [1633.8, 1480.3357142857144],
-      [1633.8, 1480.3357142857144],
-      [1633.8, 1480.3357142857144],
-      [1633.8, 1480.3357142857144],
-      [1629.9828571428573, 1476.5185714285715],
+      [1828.4742857142858, 2343.01],
+      [1855.1942857142858, 1403.9928571428572],
     ],
-    dataAttrs: {
-      'data-category': 'abnormal',
-    },
+    labelPosition: [1828.4742857142858, 2343.01],
+    lineWidth: 1.5,
   },
   {
     id: 6,
     polygon: [
-      [717.6857142857143, 1556.6785714285713],
-      [717.6857142857143, 1556.6785714285713],
-      [706.2342857142858, 1560.4957142857143],
-      [694.7828571428571, 1568.13],
-      [679.5142857142857, 1579.5814285714287],
-      [671.88, 1587.2157142857143],
-      [668.0628571428572, 1594.85],
-      [664.2457142857143, 1602.4842857142858],
-      [664.2457142857143, 1610.1185714285714],
-      [668.0628571428572, 1610.1185714285714],
-      [668.0628571428572, 1613.9357142857143],
-      [668.0628571428572, 1617.7528571428572],
-      [671.88, 1621.57],
-      [675.6971428571428, 1621.57],
-      [683.3314285714285, 1625.3871428571429],
-      [694.7828571428571, 1625.3871428571429],
-      [706.2342857142858, 1625.3871428571429],
-      [717.6857142857143, 1625.3871428571429],
-      [732.9542857142857, 1625.3871428571429],
-      [740.5885714285714, 1621.57],
-      [744.4057142857143, 1617.7528571428572],
-      [748.2228571428572, 1617.7528571428572],
-      [748.2228571428572, 1613.9357142857143],
-      [748.2228571428572, 1613.9357142857143],
-      [748.2228571428572, 1610.1185714285714],
-      [748.2228571428572, 1602.4842857142858],
-      [748.2228571428572, 1598.6671428571428],
-      [748.2228571428572, 1591.0328571428572],
-      [744.4057142857143, 1587.2157142857143],
-      [744.4057142857143, 1583.3985714285714],
-      [740.5885714285714, 1579.5814285714287],
-      [740.5885714285714, 1575.7642857142857],
-      [736.7714285714286, 1571.9471428571428],
-      [736.7714285714286, 1571.9471428571428],
-      [736.7714285714286, 1571.9471428571428],
-      [736.7714285714286, 1568.13],
-      [732.9542857142857, 1568.13],
-      [732.9542857142857, 1564.3128571428572],
-      [729.1371428571429, 1564.3128571428572],
-      [725.32, 1560.4957142857143],
-      [725.32, 1560.4957142857143],
+      [1813.2057142857143, 861.9585714285714],
+      [1328.4285714285713, 1251.307142857143],
     ],
-    dataAttrs: {
-      'data-category': 'abnormal',
-    },
+    labelPosition: [1813.2057142857143, 861.9585714285714],
+    lineWidth: 1.5,
+  },
+  {
+    id: 7,
+    polygon: [
+      [1500.2, 510.7814285714286],
+      [1171.9257142857143, 1148.2442857142858],
+    ],
+    labelPosition: [1500.2, 510.7814285714286],
+    lineWidth: 1.5,
+  },
+  {
+    id: 8,
+    polygon: [
+      [713.8685714285714, 464.9757142857143],
+      [836.0171428571429, 1312.3814285714286],
+    ],
+    labelPosition: [713.8685714285714, 464.9757142857143],
+    lineWidth: 1.5,
+  },
+  {
+    id: 9,
+    polygon: [
+      [759.6742857142857, 1472.7014285714286],
+      [1374.2342857142858, 1411.6271428571429],
+    ],
+    labelPosition: [759.6742857142857, 1472.7014285714286],
+    lineWidth: 1.5,
+  },
+  {
+    id: 10,
+    polygon: [
+      [1358.9657142857143, 1602.4842857142858],
+      [1954.44, 987.9242857142857],
+    ],
+    labelPosition: [1358.9657142857143, 1602.4842857142858],
+    lineWidth: 1.5,
+  },
+  {
+    id: 11,
+    polygon: [
+      [2004.062857142857, 148.15285714285716],
+      [2160.5657142857144, 923.0328571428571],
+    ],
+    labelPosition: [2004.062857142857, 148.15285714285716],
+    lineWidth: 1.5,
   },
 ]
